refactor draft
