+++ pyserini/search/lucene/ltr/__init__.py
@@ -20,10 +20,10 @@ from ._base import FeatureExtractor, BM25Stat, LmDirStat, DfrGl2Stat, DfrInExpB2
     AvgPooler, SumPooler, MedianPooler, MinPooler, MaxPooler, VarPooler, TfStat, TfIdfStat, NormalizedTfStat, \
     IdfStat, IcTfStat, ConfidencePooler, MaxMinRatioPooler, \
     NormalizedTfIdf, ProbalitySum, RunList, IbmModel1
-
+from ._search_msmarco import MsmarcoLtrSearcher
 __all__ = ['FeatureExtractor', 'BM25Stat', 'LmDirStat',  'DfrGl2Stat', 'DfrInExpB2Stat', 'DphStat', 'Proximity', 'TpScore', 'TpDist',
            'DocSize', 'MatchingTermCount', 'QueryLength', 'SCS', 'SumMatchingTF', 'UniqueTermCount', 'QueryCoverageRatio',
            'UnorderedSequentialPairs', 'OrderedSequentialPairs', 'UnorderedQueryPairs', 'OrderedQueryPairs',
            'AvgPooler', 'SumPooler', 'MedianPooler', 'MinPooler', 'MaxPooler', 'VarPooler', 'TfStat', 'TfIdfStat',
            'NormalizedTfStat','IdfStat', 'IcTfStat', 'ConfidencePooler', 'MaxMinRatioPooler','NormalizedTfIdf',
-            'ProbalitySum', 'RunList', 'IbmModel1']
+            'ProbalitySum', 'RunList', 'IbmModel1', 'MsmarcoLtrSearcher']
